+++ pyxpcm/__init__.py
@@ -1,12 +1,8 @@
-"""
-.. moduleauthor:: Guillaume Maze <gmaze@ifremer.fr>
-"""
-
 __version__ = "0.4.0"
 
 # Import class:
 from .pcmodel import pcm as pcm
-from .pcmodel import ds_xarray_accessor_pyXpcm as ds_pyxpcm
+from .pcmodel import ds_xarray_accessor_pyXpcm as xr_pyxpcm
 
 # Import
 # from . import stats